Renaming to dask-pytorch-ddp and dask_pytorch_ddp, repo name change still reqd
