+++ README.md
@@ -1,8 +1,8 @@
-# dask-pytorch
+# dask-pytorch-ddp
 
-<!-- ![GitHub Actions](https://github.com/saturncloud/dask-pytorch/workflows/GitHub%20Actions/badge.svg) [![PyPI Version](https://img.shields.io/pypi/v/prefect-saturn.svg)](https://pypi.org/project/prefect-saturn) -->
+<!-- ![GitHub Actions](https://github.com/saturncloud/dask-pytorch-ddp/workflows/GitHub%20Actions/badge.svg) [![PyPI Version](https://img.shields.io/pypi/v/prefect-saturn.svg)](https://pypi.org/project/prefect-saturn) -->
 
-`dask-pytorch` is a Python package that makes it easy to train PyTorch models on Dask clusters using distributed data parallel.  The intended scope of the project is
+`dask-pytorch-ddp` is a Python package that makes it easy to train PyTorch models on Dask clusters using distributed data parallel.  The intended scope of the project is
 - bootstrapping PyTorch workers on top of a Dask cluster
 - Using distributed data stores (e.g., S3) as normal PyTorch datasets
 - mechanisms for tracking and logging intermediate results, training statistics, and checkpoints.
@@ -67,13 +67,13 @@ for epoch in range(n_epochs):
 
 ## Now on Dask
 
-With dask_pytorch and PyTorch Distributed Data Parallel, we can train on multiple workers as follows:
+With dask_pytorch_ddp and PyTorch Distributed Data Parallel, we can train on multiple workers as follows:
 
 ### Loading Data
 Load the dataset from S3, and explicitly set the multiprocessing context (Dask defaults to spawn, but pytorch is generally configured to use fork)
 
 ```python
-from dask_pytorch.data import S3ImageFolder
+from dask_pytorch_ddp.data import S3ImageFolder
 
 whole_dataset = S3ImageFolder(bucket, prefix, transform=transform)
 train_loader = torch.utils.data.DataLoader(
@@ -145,11 +145,11 @@ def run_transfer_learning(bucket, prefix, samplesize, n_epochs, batch_size, num_
 
 ## How does it work?
 
-`dask-pytorch` is largely a wrapper around existing `pytorch` functionality.  `pytorch.distributed` provides infrastructure for [Distributed Data Parallel](https://pytorch.org/tutorials/intermediate/ddp_tutorial.html) (DDP).
+`dask-pytorch-ddp` is largely a wrapper around existing `pytorch` functionality.  `pytorch.distributed` provides infrastructure for [Distributed Data Parallel](https://pytorch.org/tutorials/intermediate/ddp_tutorial.html) (DDP).
 
 In DDP, you create N workers, and the 0th worker is the "master", and coordinates the synchronization of buffers and gradients.  In SGD, gradients are normally averaged between all data points in a batch.  By running batches on multiple workers, and averaging the gradients, DDP enables you to run SGD with a much bigger batch size `(N * batch_size)`
 
-`dask-pytorch` sets some environment variables to configure the "master" host and port, and then calls `init_process_group` before training, and calls `destroy_process_group` after training.  This is the same process normally done manually by the data scientist.
+`dask-pytorch-ddp` sets some environment variables to configure the "master" host and port, and then calls `init_process_group` before training, and calls `destroy_process_group` after training.  This is the same process normally done manually by the data scientist.
 
 ### Multi GPU machines
 `dask_cuda_worker` automatically rotates `CUDA_VISIBLE_DEVICES` for each worker it creates (typically one per GPU).  As a result, your PyTorch code should always start with the 0th GPU.
@@ -158,7 +158,7 @@ For example, if I have an 8 GPU machine, the 3rd worker will have `CUDA_VISIBLE_
 
 ## What else?
 
-`dask-pytorch` also implements an S3 based `ImageFolder`.  More distributed friendly datasets are planned.  `dask-pytorch`  also implements a basic results aggregation framework so that it is easy to collect training metrics across different workers.  Currently, only `DaskResultsHandler` which leverages [Dask pub-sub communication protocols][1] is implemented, but an S3 based result handler is planned.
+`dask-pytorch-ddp` also implements an S3 based `ImageFolder`.  More distributed friendly datasets are planned.  `dask-pytorch-ddp` also implements a basic results aggregation framework so that it is easy to collect training metrics across different workers.  Currently, only `DaskResultsHandler` which leverages [Dask pub-sub communication protocols][1] is implemented, but an S3 based result handler is planned.
 
 [1]:https://docs.dask.org/en/latest/futures.html#publish-subscribe
