+++ tests/test_data.py
@@ -1,12 +1,12 @@
 from unittest.mock import Mock, patch, ANY
 
 
-from dask_pytorch.data import S3ImageFolder
+from dask_pytorch_ddp.data import S3ImageFolder
 
 
 def test_image_folder_constructor():
     fake_file_list = ["d/a.jpg", "c/b.jpg"]
-    with patch("dask_pytorch.data._list_all_files", return_value=fake_file_list):
+    with patch("dask_pytorch_ddp.data._list_all_files", return_value=fake_file_list):
         fake_transform = Mock()
         fake_target_transform = Mock()
         folder = S3ImageFolder(
@@ -24,17 +24,17 @@ def test_image_folder_constructor():
 
 def test_image_folder_len():
     fake_file_list = ["d/a.jpg", "c/b.jpg"]
-    with patch("dask_pytorch.data._list_all_files", return_value=fake_file_list):
+    with patch("dask_pytorch_ddp.data._list_all_files", return_value=fake_file_list):
         folder = S3ImageFolder("fake-bucket", "fake-prefix/fake-prefix")
     assert len(folder) == 2
 
 
 def test_image_folder_getitem():
     fake_file_list = ["d/a.jpg", "c/b.jpg"]
-    with patch("dask_pytorch.data._list_all_files", return_value=fake_file_list):
+    with patch("dask_pytorch_ddp.data._list_all_files", return_value=fake_file_list):
         folder = S3ImageFolder("fake-bucket", "fake-prefix/fake-prefix")
-    with patch("dask_pytorch.data._read_s3_fileobj") as read_s3_fileobj, patch(
-        "dask_pytorch.data._load_image_obj"
+    with patch("dask_pytorch_ddp.data._read_s3_fileobj") as read_s3_fileobj, patch(
+        "dask_pytorch_ddp.data._load_image_obj"
     ) as load_image_obj:
 
         read_s3_fileobj.return_value = Mock()
